// src/types/order.ts
export interface OrderUser {
  first_name: string;
  last_name: string;
  patronymic: string;
  phone: string;
  email: string;
  payment_method: 'cash' | 'online';
}

export interface PersonInfo {
  is_adult: boolean;
  first_name: string;
  last_name: string;
  patronymic: string;
  gender: boolean;
  birthday: string;
  document_type: string;
  document_data: string;
}

export interface OrderSeat {
  coach_id: string;
  person_info: PersonInfo;
  seat_number: number;
  is_child: boolean;
  include_children_seat: boolean;
}

export interface OrderDirection {
  route_direction_id: string;
  seats: OrderSeat[];
}

export interface OrderPayload {
  user: OrderUser;
  departure: OrderDirection;
  arrival?: OrderDirection;
}